requirements.txt
fastapi
uvicorn[standard]
supabase
anthropic
openai
pydantic
pydantic-settings
python-dotenv
httpx
asyncpg
python-jose[cryptography]
passlib[bcrypt]
